c++: fix int width warning

+++ cpp/examples/ex1.cpp
@@ -2,6 +2,7 @@
 #include <string>				// std::string
 #include <sstream>				// std::stringstream
 #include <cstring> 				// strcmp()
+#include <cstddef>				// size_t
 
 #include "Huffman.hpp"		// HuffmanTreeBuilder, HuffmanEncoder, HuffmanDecoder, HuffmanTree
 
@@ -38,7 +39,7 @@ int main(int argc, char **argv)
 std::string b_vec_to_str(std::vector<bool> input)
 {
 	std::stringstream ss;
-	for (int c = 0; c < input.size(); c++) {
+	for (size_t c = 0; c < input.size(); c++) {
 		bool current = input.at(c);
 		ss << current;
 	}
